Introduce SymbolRenderer to remove _ and .erb
Previously, we used the Action View PartialRenderer to render the svg
files, which meant that they had to be prefixed with an underscore and
end in .erb.

This commit introduces a new SymbolRenderer that removes both of these
requirements so that svg files can be copied into an app and used as is.

Co-authored-by: Etienne <tiennebrb@gmail.com>