Hide symbols from DOM with display svg attribute
Previously, all_used_symbols would render a 300 by 150 empty SVG that
contains all of the symbols refernced on the page. While the SVG would
correctly not display anything on the page, it was still visible in the
DOM and hovering it would show that it was being rendered.

Initially I thought this couldn't be solved with display="none" due to
CSP policies blocking inline style. However, SVGs have a display
attribute that behaves the same as the CSS but without using CSS.
Therefore, this commit is able to take advantage of that attribute to
hide the all_used_symbols SVG without breaking any CSP.